Keep uiSubMbType[4] to values in uint8_t. Update encoder accordingly. 0x00000001, 0x00000002 0x00000004 0x00000008

+++ codec/encoder/core/inc/svc_enc_macroblock.h
@@ -47,33 +47,35 @@ namespace WelsEnc {
 // keep the most essential level pData structure be 64 Bytes, which matches cache line size; if so, the order with structure maybe negligible.
 // pls take care when modify MB structure size
 typedef struct TagMB {
-/*************************mb_layer() syntax and generated********************************/
-/*mb_layer():*/
-Mb_Type         uiMbType;       // including MB detailed partition type, number and type of reference list
-uint32_t         uiSubMbType[4]; // sub MB types
-int32_t         iMbXY;          // offset position of MB top left point based
-int16_t         iMbX;           // position of MB in horizontal axis [0..32767]
-int16_t         iMbY;           // position of MB in vertical axis [0..32767]
+  /*************************mb_layer() syntax and generated********************************/
+  /*mb_layer():*/
+  Mb_Type         uiMbType;       // including MB detailed partition type, number and type of reference list
+  uint8_t         uiSubMbType[4]; // sub MB types
+  int32_t         iMbXY;          // offset position of MB top left point based
+  int16_t         iMbX;           // position of MB in horizontal axis [0..32767]
+  int16_t         iMbY;           // position of MB in vertical axis [0..32767]
 
-uint8_t         uiNeighborAvail;        // avail && same_slice: LEFT_MB_POS:0x01, TOP_MB_POS:0x02, TOPRIGHT_MB_POS = 0x04 ,TOPLEFT_MB_POS = 0x08;
-uint8_t         uiCbp;
+  uint8_t
+  uiNeighborAvail;        // avail && same_slice: LEFT_MB_POS:0x01, TOP_MB_POS:0x02, TOPRIGHT_MB_POS = 0x04 ,TOPLEFT_MB_POS = 0x08;
+  uint8_t         uiCbp;
 
-SMVUnitXY*      sMv;
-int8_t*         pRefIndex;
+  SMVUnitXY*      sMv;
+  int8_t*         pRefIndex;
 
-int32_t*        pSadCost;          // mb sad. set to 0 for intra mb
-int8_t*         pIntra4x4PredMode; // [MB_BLOCK4x4_NUM]
-int8_t*         pNonZeroCount;     // [MB_LUMA_CHROMA_BLOCK4x4_NUM]
+  int32_t*        pSadCost;          // mb sad. set to 0 for intra mb
+  int8_t*         pIntra4x4PredMode; // [MB_BLOCK4x4_NUM]
+  int8_t*         pNonZeroCount;     // [MB_LUMA_CHROMA_BLOCK4x4_NUM]
 
-SMVUnitXY       sP16x16Mv;
+  SMVUnitXY       sP16x16Mv;
 
-uint8_t         uiLumaQp;       // uiLumaQp: pPps->iInitialQp + sSliceHeader->delta_qp + mb->dquant.
-uint8_t         uiChromaQp;
-uint16_t        uiSliceIdc;     // 2^16=65536 > MaxFS(36864) of level 5.1; AVC: iFirstMbInSlice?; SVC: (iFirstMbInSlice << 7) | ((uiDependencyId << 4) | uiQualityId);
-uint32_t        uiChromPredMode;
-int32_t         iLumaDQp;
-SMVUnitXY       sMvd[MB_BLOCK4x4_NUM]; //only for CABAC writing; storage structure the same as sMv, in 4x4 scan order.
-int32_t         iCbpDc;
+  uint8_t         uiLumaQp;       // uiLumaQp: pPps->iInitialQp + sSliceHeader->delta_qp + mb->dquant.
+  uint8_t         uiChromaQp;
+  uint16_t
+  uiSliceIdc;     // 2^16=65536 > MaxFS(36864) of level 5.1; AVC: iFirstMbInSlice?; SVC: (iFirstMbInSlice << 7) | ((uiDependencyId << 4) | uiQualityId);
+  uint32_t        uiChromPredMode;
+  int32_t         iLumaDQp;
+  SMVUnitXY       sMvd[MB_BLOCK4x4_NUM]; //only for CABAC writing; storage structure the same as sMv, in 4x4 scan order.
+  int32_t         iCbpDc;
 //uint8_t         reserved_filling_bytes[1];      // not deleting this line for further changes of this structure. filling bytes reserved to make structure aligned with 4 bytes, higher cache hit on less structure size by 2 cache lines( 2 * 64 bytes) once hit
 } SMB, *PMb;
